Cleanup split navigators code
