+++ src/libs/Navigation/AppNavigator/Navigators/ReportsSplitNavigator.tsx
@@ -15,7 +15,7 @@ import SCREENS from '@src/SCREENS';
 import type ReactComponentModule from '@src/types/utils/ReactComponentModule';
 
 const loadReportScreen = () => require<ReactComponentModule>('@pages/home/ReportScreen').default;
-const loadSidebarScreen = () => require<ReactComponentModule>('@pages/home/sidebar/SidebarScreen').default;
+const loadSidebarScreen = () => require<ReactComponentModule>('@pages/home/sidebar/BaseSidebarScreen').default;
 
 const Split = createSplitNavigator<ReportsSplitNavigatorParamList>();
