+++ src/libs/Navigation/AppNavigator/Navigators/WorkspaceSplitNavigator.tsx
@@ -47,7 +47,7 @@ function WorkspaceSplitNavigator({route, navigation}: PlatformStackScreenProps<A
                 return;
             }
 
-            // We want ot set animation after mounting so it will animate on going UP to the settings split.
+            // We want to set animation after mounting so it will animate on going UP to the settings split.
             navigation.setOptions({animation: Animations.SLIDE_FROM_RIGHT});
         });
