+++ src/pages/home/sidebar/BottomTabBarFloatingActionButton/index.tsx
@@ -1,5 +1,5 @@
 import React, {useCallback, useRef} from 'react';
-import FloatingActionButtonAndPopover from '@pages/home/sidebar/SidebarScreen/FloatingActionButtonAndPopover';
+import FloatingActionButtonAndPopover from '@pages/home/sidebar/FloatingActionButtonAndPopover';
 import type FloatingActionButtonPopoverMenuRef from './types';
 
 type BottomTabBarFloatingActionButtonProps = {
