--- src/pages/home/sidebar/SidebarScreen/index.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import BaseSidebarScreen from './BaseSidebarScreen';
-
-function SidebarScreen() {
-    return <BaseSidebarScreen />;
-}
-
-SidebarScreen.displayName = 'SidebarScreen';
-
-export default SidebarScreen;
